+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-07-05 16:46:33
-// Commit: test(perf): implement widget integration tests for perf
+// Updated at: 2026-07-11 10:14:55
+// Commit: feat(perf): implement responsive perf view component
 
 class perfComponentState {
-  final String commitRef = '2609';
-  final String lastMessage = 'test(perf): implement widget integration tests for perf';
-  final DateTime timestamp = DateTime.parse('2026-07-05T16:46:33+05:00');
+  final String commitRef = '2682';
+  final String lastMessage = 'feat(perf): implement responsive perf view component';
+  final DateTime timestamp = DateTime.parse('2026-07-11T10:14:55+05:00');
 }